*** empty log message ***

+++ bmake/t3e/petscfix.h
@@ -84,7 +84,7 @@ extern void   perror(const char *);
 extern double atof(const char *);
 extern void   free(void *);
 extern void   *malloc(long unsigned int );
-extern int    readlink(const char *,char *,int);
+extern int    readlink(const char *,char *,size_t);
 }
 
 #else
